Remove `IteratesControllers` trait
While `IteratesControllers` was designed with several use cases in mind, it's only used in a single test class for the foreseeable future.  This PR moves the `getControllers` method from the trait to the single class where it's used to reduce clutter.
